Add documentation and implement GitHub workflow for deployment
The commit introduces deployment configuration files for enabling GitHub actions for deploying the documentation. The vital component being a workflow script `.github/workflows/deploy.yml` which has build and deployment tasks. This script automates the build of the documentation and its deployment on every push to the main branch or when manually triggered from the Actions tab on GitHub. The changes also reflect the addition of several new documentation files, modifications to the .gitignore file, and the creation of .idea configuration files.